Moved the AutoConst transform to a separate file to prep for DriveBase consumption
