+++ tsconfig.json
@@ -8,7 +8,8 @@
     "allowJs": false,
     "esModuleInterop": true,
     "noEmit": true,
-    "pretty": true
+    "pretty": true,
+    "skipLibCheck": true
   },
   "include": ["scripts/*.ts", "scripts/helpers/*.ts"]
 }
